Feature/wiki rockets

+++ app/src/main/java/com/github/emilg1101/spacex/data/spacexapi/pojo/response/Engines.kt
@@ -12,5 +12,5 @@ data class Engines(
     @SerializedName("propellant_2") val propellant2: String,
     @SerializedName("thrust_sea_level") val thrustSeaLevel: ThrustSeaLevel,
     @SerializedName("thrust_vacuum") val thrustVacuum: ThrustVacuum,
-    @SerializedName("thrust_to_weight") val thrustToWeight: Int
+    @SerializedName("thrust_to_weight") val thrustToWeight: Double
 )
